refactor: 디렉렉터리 구조 변경
